requirements.txt
flask 
pymongo 
dnspython